% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Qspline_signal.R
\name{generate_Qspline_signal}
\alias{generate_Qspline_signal}
\title{Generate a Quadratic Spline Signal with Additive Noise}
\usage{
generate_Qspline_signal(result, segments, noise_sd = 1)
}
\arguments{
\item{result}{List output from `generate_Qsplines`, containing vectors `p`, `v`, and `a`
for the position, velocity, and acceleration of each segment at its start.}

\item{segments}{Numeric vector of segment lengths used to generate the spline.}

\item{noise_sd}{Standard deviation of additive Gaussian noise (default 0).}
}
\value{
A numeric vector representing the continuous spline signal with noise.
}
\description{
Generate a Quadratic Spline Signal with Additive Noise
}
\examples{
segments <- c(30, 20, 40)
accelerations <- c(1,-1,1)
result <- generate_Qsplines(segments, accelerations)
signal <- generate_Qspline_signal(result, segments, noise_sd = 0.01)
plot(signal, type = "l")
}
